+++ RELEASENOTES.md
@@ -2,6 +2,8 @@ These are the release notes for the [mediawiki-api-base](README.md).
 
 ## Version 0.2 (under development)
 
+* Unsuccessful logins now throw a UsageException with extra details
+
 ## Version 0.1.2 (25 May 2014)
 
 * Fix issue where API tokens were not returned